cli: add `delete signer` command

+++ crates/coinstr-sdk/src/client.rs
@@ -646,14 +646,16 @@ impl Coinstr {
         signer_id: EventId,
         _timeout: Option<Duration>,
     ) -> Result<(), Error> {
-        let keys = self.client.keys();
+        if signer_id != EventId::all_zeros() {
+            let keys = self.client.keys();
 
-        // TODO: delete shared signers and notify the interested users (tag pubkey in deletion event)
+            // TODO: delete shared signers and notify the interested users (tag pubkey in deletion event)
 
-        let event = EventBuilder::new(Kind::EventDeletion, "", &[]).to_event(&keys)?;
-        self.send_event(event).await?;
+            let event = EventBuilder::new(Kind::EventDeletion, "", &[]).to_event(&keys)?;
+            self.send_event(event).await?;
 
-        self.db.delete_signer(signer_id)?;
+            self.db.delete_signer(signer_id)?;
+        }
 
         Ok(())
     }
